pyproject.toml
[project]
name = "rpi-gardener"
version = "0.1.0"
description = "Raspberry Pi gardening monitor"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite",
    "python-dotenv",
    "redis>=5.0",
    "starlette",
    "uvicorn[standard]",
]

[project.optional-dependencies]
hardware = [
    "adafruit-circuitpython-dht",
    "adafruit-circuitpython-ssd1306",
    "pillow",
    "aioserial",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "mypy>=1.0",
    "ruff>=0.8",
]
micropython = [
    "mpremote",
    "rshell",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rpi"]

[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
disable_error_code = [
    "arg-type",
    "return-value",
    "assignment",
    "union-attr",
    "no-any-return",
    "literal-required",
    "index",
]

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["no-untyped-def", "no-untyped-call", "arg-type"]

[tool.ruff]
line-length = 79
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by ruff format)
    "E402",   # module level import not at top (needed for type aliases, test mocking)
    "SIM102", # nested if statements (sometimes clearer than combined)
    "SIM105", # contextlib.suppress (try/except more explicit)
    "SIM108", # ternary operator (if/else often clearer)
]

[tool.ruff.lint.isort]
known-first-party = ["rpi"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
